back/requirements.txt
flask==2.3.3
Flask-Cors==4.0.0
llama-cpp-python==0.2.13
gguf==0.4.5
transformers==4.35.0
torch==2.1.0
accelerate==0.24.1
einops==0.7.0
